Improve support for SQLAlchemy 1.4.x
It's necessary to pass future=True when calling create_engine() in
SQLAlchemy 1.4.x in order for nested transactions to work as they do
in SQLAlchemy 2.0.x.

On SQLAlchemy 1.3.x, the future=True flag isn't supported, but it seems
that nested transactions work nevertheless. This is surprising, but the
test suite passes...